feat: Add getTypeList function for converting dynamic data to a list of objects
This code change adds a new function called `getTypeList` that can be used to convert dynamic data into a list of objects of a specified type. It takes in the dynamic data and a function `fromJson` that converts a JSON object into an object of the specified type. The function returns a list of objects of type T.

Example usage:
```dart
final nearbyPharmacies = getTypeList<NearbyPharmacies>(
  response.data,
  NearbyPharmacies.fromJson,
);