feat: Update easix package description and version
This code change updates the description and version of the easix package in the pubspec.yaml file. The description now includes information about the multiple handy functions and extensions provided by the package. The version has been incremented to 1.1.0 to reflect the added functionality. This update improves the clarity and accuracy of the package information.